+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 * add tzcompose alpha release
+* return full BlockHeaderLogEntry in BlockObserver callback
 * dedicated logger instance per client
 * block observer multiple subscriptions for the same op hash
 * block observer support block subscriptions (use zero op hash)